CoreComm: move to contikimote package
This class no longer needs to be exposed
to the world.
