+++ java/org/contikios/cooja/contikimote/ContikiMoteType.java
@@ -48,7 +48,6 @@ import java.util.regex.Pattern;
 import org.contikios.cooja.AbstractionLevelDescription;
 import org.contikios.cooja.ClassDescription;
 import org.contikios.cooja.Cooja;
-import org.contikios.cooja.CoreComm;
 import org.contikios.cooja.Mote;
 import org.contikios.cooja.MoteInterface;
 import org.contikios.cooja.ProjectConfig;
